Added Toggle class.  Removed index.js
