+++ Murach's JavaScript and JQuery/Chapter 06/FAQs/index.html
@@ -6,7 +6,7 @@
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>FAQs</title>
     <link rel="stylesheet" type="text/css" href="index.css">
-    <script src="index.js"></script>
+    
 </head>
 <body>
     <main id="faqs">
@@ -33,5 +33,8 @@
             </ul>
         </div>
     </main>
+
+    <script type="module" src="toggle.js"></script>
+    <script type="module" src="main.js"></script>
 </body>
 </html>
