+++ Murach's JavaScript and JQuery/Chapter 06/FAQs/main.js
@@ -1,0 +1,7 @@
+import Toggle from './toggle.js'
+
+window.onload = function() {
+    var toggle = new Toggle();
+
+    toggle.run();
+}
